Remove deploy key usage and instead use temporary SSH key

+++ tests/test_buildrunner_files.py
@@ -26,29 +26,32 @@ serial_test_files = [
 @pytest.fixture(autouse=True, scope="session")
 def setup_buildrunner_test_ssh_key():
     key_file_path = Path(TEST_SSH_KEY_FILE)
-    key_file_path.unlink(missing_ok=True)
-    subprocess.run(
-        [
-            "ssh-keygen",
-            "-t",
-            "ecdsa",
-            "-m",
-            "PEM",
-            "-N",
-            "",
-            "-f",
-            TEST_SSH_KEY_FILE,
-        ],
-        check=True,
-    )
+    cleanup_key_file = False
+    pub_key_file_path = Path(f"{TEST_SSH_KEY_FILE}.pub")
+    if not key_file_path.exists():
+        subprocess.run(
+            [
+                "ssh-keygen",
+                "-t",
+                "ecdsa",
+                "-m",
+                "PEM",
+                "-N",
+                "",
+                "-f",
+                TEST_SSH_KEY_FILE,
+            ],
+            check=True,
+        )
+        cleanup_key_file = True
     # Set the public key in an environment variable to use in the test buildrunner files
-    os.environ["BUILDRUNNER_TEST_SSH_PUB_KEY"] = (
-        Path(f"{TEST_SSH_KEY_FILE}.pub").read_text().strip()
-    )
+    os.environ["BUILDRUNNER_TEST_SSH_PUB_KEY"] = pub_key_file_path.read_text().strip()
     yield
     # Cleanup
     del os.environ["BUILDRUNNER_TEST_SSH_PUB_KEY"]
-    key_file_path.unlink()
+    if cleanup_key_file:
+        key_file_path.unlink()
+        pub_key_file_path.unlink()
 
 
 def _get_test_args(file_name: str) -> Optional[List[str]]:
